NIAID-215: API: - EntitySearchResult: changed offset type from int to long - DefaultDocumentSearchImpl: can now do count query only - QueryExpression: fixed negation handling - AbstractFilterBuilder: added refValues - DocumentSearchBuilder: added negation - DocumentSearchBuilder: overloaded onlyForUser can now control ownership and visibility checks
REST:
- refactored chunks of DefaultLiveTableSearchImpl into new classes: new LiveTableGenerator (and impl) and LiveTableFacade

+++ phenotips/entities-search/api/src/main/java/org/phenotips/data/api/EntitySearchResult.java
@@ -24,7 +24,7 @@ public class EntitySearchResult<T>
 {
     private long totalRows;
 
-    private int offset;
+    private long offset;
 
     private List<T> items = new LinkedList<>();
 
@@ -66,7 +66,7 @@ public class EntitySearchResult<T>
      *
      * @return offset
      */
-    public int getOffset()
+    public long getOffset()
     {
         return this.offset;
     }
@@ -77,7 +77,7 @@ public class EntitySearchResult<T>
      * @param offset offset to set
      * @return this object
      */
-    public EntitySearchResult<T> setOffset(int offset)
+    public EntitySearchResult<T> setOffset(long offset)
     {
         this.offset = offset;
         return this;
